Incorporate review feedback

+++ src/block/compress.rs
@@ -387,7 +387,7 @@ pub(crate) fn compress_internal<T: HashTable, const USE_DICT: bool>(
         assert!(input_stream_offset
             .checked_add(input.len())
             .and_then(|i| i.checked_add(ext_dict.len()))
-            .map_or(false, |i| i <= usize::MAX));
+            .map_or(false, |i| i <= isize::MAX as usize));
     }
     if output.capacity() - output.pos() < get_maximum_output_size(input.len() - input_pos) {
         return Err(CompressError::OutputTooSmall);
